fix: strengthen confidence prompt instructions so model returns the field

+++ companion/src/analysis/responseSchema.ts
@@ -9,7 +9,7 @@ export const deltaSchema = z.object({
   findings: z.array(z.object({
     id: z.string().min(1),
     severity: severity.catch("Medium"),
-    confidence: z.number().int().min(0).max(100).optional().catch(undefined),
+    confidence: z.number().min(0).max(100).optional().catch(undefined),
     title: z.string().min(1),
     description: z.string(),
     relatedIocs: z.array(z.string()),
